pyproject.toml
[tool.poetry]
name = "data-mosher"
version = "0.1.0"
description = "This software takes in videos and performs a `glitch art` effect called datamoshing"
authors = ["Mia Abbott <mia.abbott96@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "3.13"
opencv-python = "^4.10.0.84"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
